# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OurLog API",
    "description": "OurLog 다이어리 애플리케이션 API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "개발 서버"
    }
  ],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "summary": "로그인",
        "tags": ["인증"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "summary": "회원가입",
        "tags": ["인증"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "회원가입 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "로그아웃",
        "tags": ["인증"],
        "responses": {
          "200": {
            "description": "로그아웃 성공"
          }
        }
      }
    },
    "/api/v1/diaries": {
      "get": {
        "summary": "다이어리 목록 조회",
        "tags": ["다이어리"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "다이어리 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "다이어리 생성",
        "tags": ["다이어리"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiaryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "다이어리 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diaries/{diaryId}": {
      "get": {
        "summary": "다이어리 상세 조회",
        "tags": ["다이어리"],
        "parameters": [
          {
            "name": "diaryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "다이어리 상세 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diary"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "다이어리 수정",
        "tags": ["다이어리"],
        "parameters": [
          {
            "name": "diaryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiaryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "다이어리 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "다이어리 삭제",
        "tags": ["다이어리"],
        "parameters": [
          {
            "name": "diaryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "다이어리 삭제 성공"
          }
        }
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "summary": "사용자 프로필 조회",
        "tags": ["사용자"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 프로필 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/profile": {
      "put": {
        "summary": "프로필 수정",
        "tags": ["사용자"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "프로필 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "nickname": {
            "type": "string"
          },
          "profileImageUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["id", "email", "nickname"]
      },
      "Diary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "title", "content", "createdAt", "updatedAt"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["email", "password"]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          }
        },
        "required": ["email", "password", "nickname"]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": ["accessToken", "user"]
      },
      "CreateDiaryRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": ["title", "content"]
      },
      "UpdateDiaryRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string"
          },
          "profileImageUrl": {
            "type": "string"
          }
        }
      },
      "DiaryListResponse": {
        "type": "object",
        "properties": {
          "diaries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Diary"
            }
          },
          "totalCount": {
            "type": "integer"
          },
          "currentPage": {
            "type": "integer"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        },
        "required": ["diaries", "totalCount", "currentPage", "hasNextPage"]
      }
    }
  }
} 